Speed up serial line reads
A loop cycle in esphome happens once about every 16ms and an opener command on serial is transmitted in about 16ms as well,
so by the time a byte can be read on the serial line the whole command can probably be read. This change reads the whole command
at once (or as much as possible) instead of a byte on every loop cycle.